chore: move slack service and use it to alert outdated push schedules
