+++ app-server/subprojects/bounded_context/notification/application/build.gradle.kts
@@ -1,3 +1,4 @@
 dependencies {
     implementation(projects.boundedContext.user.application)
+    implementation(projects.boundedContext.misc.application)
 }
