+++ app-server/subprojects/cross_cutting_concern/infra/spring_web/src/integrationTest/resources/application.yaml
@@ -70,6 +70,7 @@ scc:
   slack:
     token: test
     channel:
+      alert: "#scc-server-alert"
       reportAccessibility: "#scc-accessibility-report-test"
 
   open-data:
