+++ app-server/subprojects/cross_cutting_concern/test/spring_it/src/main/resources/application.yaml
@@ -76,6 +76,7 @@ scc:
   slack:
     token: test
     channel:
+      alert: "#scc-server-alert"
       reportAccessibility: "#scc-accessibility-report-test"
 
   open-data:
