+++ app-server/subprojects/deploying_apps/scc_server/src/main/resources/application.yaml
@@ -95,6 +95,7 @@ scc:
   slack:
     token: test
     channel:
+      alert: "#scc-server-alert"
       reportAccessibility: "#scc-accessibility-report-test"
 
   open-data:
